refactor core comicarchive classes in its own package comicapi
