+++ comicapi/UnRAR2/UnRARDLL/whatsnew.txt
@@ -1,0 +1,80 @@
+List of unrar.dll API changes. We do not include performance and reliability
+improvements into this list, but this library and RAR/UnRAR tools share
+the same source code. So the latest version of unrar.dll usually contains
+same decompression algorithm changes as the latest UnRAR version.
+============================================================================
+
+--  18 January 2008
+
+all LONG parameters of CallbackProc function were changed
+to LPARAM type for 64 bit mode compatibility.
+
+
+--  12 December 2007
+
+Added new RAR_OM_LIST_INCSPLIT open mode for function RAROpenArchive.
+
+
+--  14 August 2007
+
+Added NoCrypt\unrar_nocrypt.dll without decryption code for those
+applications where presence of encryption or decryption code is not
+allowed because of legal restrictions.
+
+
+--  14 December 2006
+
+Added ERAR_MISSING_PASSWORD error type. This error is returned
+if empty password is specified for encrypted file.
+
+
+--  12 June 2003
+
+Added RARProcessFileW function, Unicode version of RARProcessFile
+
+
+--  9 August 2002
+
+Added RAROpenArchiveEx function allowing to specify Unicode archive
+name and get archive flags.
+
+
+--  24 January 2002
+
+Added RARReadHeaderEx function allowing to read Unicode file names
+and 64 bit file sizes.
+
+
+--  23 January 2002
+
+Added ERAR_UNKNOWN error type (it is used for all errors which
+do not have special ERAR code yet) and UCM_NEEDPASSWORD callback
+message.
+
+Unrar.dll automatically opens all next volumes not only when extracting,
+but also in RAR_OM_LIST mode.
+
+
+--  27 November 2001
+
+RARSetChangeVolProc and RARSetProcessDataProc are replaced by
+the single callback function installed with RARSetCallback.
+Unlike old style callbacks, the new function accepts the user defined
+parameter. Unrar.dll still supports RARSetChangeVolProc and
+RARSetProcessDataProc for compatibility purposes, but if you write
+a new application, better use RARSetCallback.
+
+File comments support is not implemented in the new DLL version yet.
+Now CmtState is always 0.
+
+
+--  13 August 2001
+
+Added RARGetDllVersion function, so you may distinguish old unrar.dll,
+which used C style callback functions and the new one with PASCAL callbacks.
+
+
+--  10 May 2001
+
+Callback functions in RARSetChangeVolProc and RARSetProcessDataProc
+use PASCAL style call convention now.
