+++ comicapi/UnRAR2/UnRARDLL/x64/readme.txt
@@ -1,0 +1,1 @@
+This is x64 version of unrar.dll.
