--- comictaggerlib/UnRAR2/UnRARDLL/x64/readme.txt
@@ -1,1 +1,0 @@
-This is x64 version of unrar.dll.
